+++ sodalite/ui/app.py
@@ -1,12 +1,15 @@
+import curses
 import logging
 import os
+import threading
+import time
 
 import npyscreen
 
 from ui import theme
 from ui.control import MainControl
 from ui.entryview import EntrySplitter
-from ui.hookpane import HookBox
+from ui.footerbox import NotifyBox, HookBox
 
 logger = logging.getLogger(__name__)
 
@@ -21,7 +24,7 @@ class App(npyscreen.NPSAppManaged):
 
     def onStart(self):
         npyscreen.setTheme(theme.Theme)
-        self.addForm('MAIN', MainForm)
+        self.main = self.addForm('MAIN', MainForm)
 
 
 class MainForm(npyscreen.FormBaseNew):
@@ -33,6 +36,7 @@ class MainForm(npyscreen.FormBaseNew):
 
     def __init__(self, *args, **keywords):
         self.splitter = None
+        self.notify_lock = threading.Lock()
         super().__init__(*args, **keywords)
 
     def create(self):
@@ -45,7 +49,25 @@ class MainForm(npyscreen.FormBaseNew):
 
         self.main_control = MainControl(self)
         self.hookpane = self.add(HookBox, main_control=self.main_control)
+        self.notifypane = self.add(NotifyBox, main_control=self.main_control)
         self.splitter = self.add(EntrySplitter, main_control=self.main_control)
 
         self.min_c = 28
         self.min_l = 10
+
+    def notify(self, message: str, duration: float, attr=curses.A_NORMAL):
+        thread = threading.Thread(target=self._notify, args=(message, duration, attr,))
+        thread.start()
+
+    def _notify(self, message: str, duration: float, attr: int):
+        self.hookpane.display()
+        self.notifypane.display()
+        self.notify_lock.acquire()
+        try:
+            logger.info(f"Show notification: '{message}'")
+            self.notifypane.notify(message, attr)
+            self.notifypane.display()
+            time.sleep(duration)
+            self.hookpane.display()
+        finally:
+            self.notify_lock.release()